Add `--auto-install` flag to phx.new generator
